Projects/Automation/databricks_runner/requirements.txt
databricks-sql-connector>=3.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
anthropic>=0.25.0
tabulate>=0.9.0
pyarrow>=14.0.0
